feat: hitsave can now be used as a cli tool without 'python -m'
